# docs/openapi.yaml
openapi: 3.0.0
info:
  title: ARP Guard API
  description: API for ARP Guard - Network Security and Device Management
  version: 1.0.0
  contact:
    name: ARP Guard Team
    email: team@arpguard.io
    url: https://arpguard.io
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.arpguard.io
    description: Production server
  - url: https://api.arpguard.io/v1
    description: Production server (explicit version)
  - url: https://staging.arpguard.io/api
    description: Staging server
...
paths:
...
  /devices:
    get:
      summary: List all devices
      description: Get a list of all devices
      tags:
        - Devices
      security:
        - apiKey: []
      parameters:
        - name: page
          in: query
          description: Page number
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          description: Number of items per page
          schema:
            type: integer
            default: 50
            maximum: 100
      responses:
        '200':
          description: List of devices
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Device'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/RateLimited'
    
    post:
      summary: Create a new device
      description: Create a new device in the system
      tags:
        - Devices
      security:
        - apiKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeviceCreate'
      responses:
        '201':
          description: Device created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Device'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          description: Device with the same MAC address already exists
        '429':
          $ref: '#/components/responses/RateLimited'
...
components:
  securitySchemes:
    apiKey:
      type: apiKey
      in: header
      name: X-API-Key
    apiKeySecret:
      type: apiKey
      in: header
      name: X-API-Secret
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    RateLimited:
      description: Rate limit exceeded
      headers:
        X-RateLimit-Limit:
          description: Rate limit ceiling for the given endpoint
          schema:
            type: integer
        X-RateLimit-Remaining:
          description: Number of requests left for the time window
          schema:
            type: integer
        X-RateLimit-Reset:
          description: Time at which the rate limit resets in UTC epoch seconds
          schema:
            type: integer
        X-RateLimit-Retry-After:
          description: Seconds to wait before retrying the request
          schema:
            type: integer
        X-RateLimit-Type:
          description: Type of rate limit (IP, user, token, or endpoint)
          schema:
            type: string
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/Error'
              - type: object
                properties:
                  retry_after:
                    type: integer
                    description: Seconds to wait before retrying the request
  schemas:
    Device:
      type: object
      properties:
        id:
          type: string
          description: Unique identifier for the device
        mac_address:
          type: string
          description: MAC address of the device
          pattern: '^([0-9A-Fa-f]{2}[:-]){5}([0-9A-Fa-f]{2})$'
        ip_address:
          type: string
          description: IP address of the device
          format: ipv4
        hostname:
          type: string
          description: Hostname of the device
        last_seen:
          type: string
          format: date-time
          description: Timestamp when the device was last seen
        first_seen:
          type: string
          format: date-time
          description: Timestamp when the device was first seen
        attributes:
          type: object
          description: Additional custom attributes
        created_at:
          type: string
          format: date-time
          description: Timestamp when the device was created
        updated_at:
          type: string
          format: date-time
          description: Timestamp when the device was last updated
      required:
        - id
        - mac_address
    DeviceCreate:
      type: object
      properties:
        mac_address:
          type: string
          description: MAC address of the device
          pattern: '^([0-9A-Fa-f]{2}[:-]){5}([0-9A-Fa-f]{2})$'
        ip_address:
          type: string
          description: IP address of the device
          format: ipv4
        hostname:
          type: string
          description: Hostname of the device
        attributes:
          type: object
          description: Additional custom attributes
      required:
        - mac_address
...
    Error:
      type: object
      properties:
        status:
          type: integer
          description: HTTP status code
        code:
          type: string
          description: Error code
        message:
          type: string
          description: Error message
      required:
        - status
        - message 